Make libbacktrace an optional dependency
libbacktrace is not available in some distributions where Extension
Manager is available as a native package, e.g. Gentoo [1], Debian [2].
On Debian, the native package maintainers have had to apply a downstream
patch [2] that removes and masks all code interfacing with libbacktrace.

However, when libbacktrace is available, the backtraces it helps to
collect in the event of an app crash might still be useful, so removing
the libbacktrace dependency completely is not the best solution for all
cases, i.e. not a good change to be submitted to the Extension Manager
upstream.

Making Extension Manager's dependency on libbacktrace optional is
better: it allows both these distributions to continue shipping this app
without downstream patching, and users who install this app via Flathub
or other distributions that support libbacktrace to report crashes with
helpful backtraces.

This commit introduces a new 'backtrace' Meson boolean option, which
controls whether Extension Manager is to be built with libbacktrace.  It
is set to 'true' by default, so existing build pipelines, e.g. the
first-party Flatpak builds for Flathub, behave the same as before.
Folks who wish to build Extension Manager without libbacktrace can do it
by setting '-Dbacktrace=false' in Meson command-line options.

[1]: https://gitweb.gentoo.org/repo/gentoo.git/commit/?id=c99e8159e6cb8705227aab759359c158f52b64bd
[2]: https://salsa.debian.org/gnome-team/gnome-shell-extension-manager/-/commit/a7cccc5486ebeca2597aeab2ea5d6c74b17488b0

Signed-off-by: Yuan Liao <liaoyuan@gmail.com>