flatpak: Have stable and devel manifests in tree
With the master version of the runtime we can test the changes in the
development versions of the libraries beforehand.

This also removes `webp-pixbuf-loader` from the modules in the devel
manifest, as it will be included in the platform for GNOME 46.
See https://gitlab.gnome.org/GNOME/gnome-build-meta/-/merge_requests/2507