upgrade-assistant: Fix GNOME version calculation
With the new year, two new versions of GNOME have been added that have not
yet been released. To fix this, since the calculation based on the release
date of GNOME 40 always results in the first version to be released in
March of that year, which is even, then:
- If the month is prior to March, subtract one from the approximation,
leaving the odd-numbered version released in September of the previous
year.
- If the month is between March and September, we keep the value.
- If the month is September or later, we add one to get the odd numbered
version of the year.