build: Use version suffix instead of name suffix
Over the last cycles, apps added to GNOME core are not using the suffix in
their names for development versions. Instead, they are adding a suffix to
the app version following the running commit. Since it simplifies build
files a bit and translators work, let's go that route as well.