detail-view: Add independent breakpoint
With the recent addition of extension's icon, the content at the top does
not look good on smaller screens.

Since we require more setters for improving this situtation, relying on
window's breakpoint will make the code harder to follow in the future.
Setting a separate breakpoint for this view solves it.