screenshot-view: Fix zoom shortcuts
When zooming out to the minimum, keyboard shortcuts suddenly stop working.

Adding the shortcut to the widget_class does not work properly here, even
thought the widget tree is mostly the same as detail-view.

Setting the scope property to `managed`, instead of `local` that behaves
the same as it does now, does the job.