extension-row: Use AdwActionRow for description
This replaces the GtkGrid with a series of AdwActionRow for the
extension description, version, and error message. This makes the design
more modern and in line with the rest of the app and other GNOME apps,
and makes the app fit on smartphones.

Fixes https://github.com/mjakeman/extension-manager/issues/767

Signed-off-by: Adrien Plazas <aplazas@gnome.org>