install-button: Disable during installation
Introduces a new installing state that is set before activating
`ext.install` action. To deal with cancellations, a new `install-status`
signal is added to ExmManager which is emitted whenever it happens
andreverts the button state to the last one it had, either default or
unsupported.

ExmSearchRow drops `is-installed` and `is-supported` properties in favor
of a new `manager` property to connect to ExmManager's `install-status
signal.