install-button: Improve adaptability in ExmSearchRow
ExmInstallButton now has a GtkStack as a child with a GtkImage page and
a GtkLabel one. A new property, icon-only, is bound to the former.

Adds an AdwBreakpoint to ExmBrowsePage to dynamically manage layout
properties. The breakpoint updates ExmSearchRow's new compact
property, which is bound to the icon-only property of ExmInstallButton.