installed-page: Save extension's toggle focus on state changes
The items-changed signal is now emitted without adding or removing items
to the model from ExmManager and is connected after binding the manager
in ExmInstalledPage from where the callback saves, emits items-changed
adding and removing as done before in ExmManager, and restores focus.

This approach allows us to keep binding GtkListBox to the extensions
model and not having to manage the changes manually.

Fix https://github.com/mjakeman/extension-manager/issues/220