Bring back supported versions section as a dialog
Drops ExmInfoBar in favor of a GtkListBox. Now that the version row has
to indicate that it opens another view with go-next-symbolic, we have
the perfect excuse to move the subtitles to the row's suffix and get rid
of the hacky CSS.

We can revisit this in the future with a more friendly widget for
horizontal layouts.