detail-view: Hide "Show All Reviews" button when unneeded
We retrieve reviews/comments in the details view and the comment dialog
using different queries. Because we do not know the total count until
querying all of them in the dialog, we have to display the button for
extensions with exactly five or more reviews/comments, rather than
strictly more than five.

Also rewords strings containing comment or review to use both terms as
they are both possible and adds pill style to the button.

Close https://github.com/mjakeman/extension-manager/issues/781