fix(ExmBrowsePage): Allow empty ordering
By default, results are sorted by descending popularity when the query
is empty and no order is selected. This is invisible to the user, as
results match the name of the extensions better when no order is set,
but worsen when the query is empty.

Fix https://github.com/mjakeman/extension-manager/issues/883